refactor: rename to `context.rs`
